scikit-learn/regression-mlops/requirements.txt
mlflow~=2.4.1
pandas~=2.0.2
plac~=1.3.5
pathlib~=1.0.1
scikit-learn~=1.2.2
PyYAML~=6.0